backend/requirements.txt
django==6.0.1
djangorestframework
django-cors-headers
qrcode
Pillow
